Remove point-to-point from results type
